refactor: separate repository subpackage from domain package
