+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -4,7 +4,7 @@ import java.util.stream.Collectors;
 import kitchenpos.application.dto.request.CreateMenuGroupDto;
 import kitchenpos.application.dto.response.MenuGroupDto;
 import kitchenpos.domain.menu.MenuGroup;
-import kitchenpos.domain.menu.repository.MenuGroupRepository;
+import kitchenpos.repository.menu.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
