+++ src/main/java/kitchenpos/application/OrderService.java
@@ -7,13 +7,13 @@ import kitchenpos.application.dto.response.OrderDto;
 import kitchenpos.application.dto.request.UpdateOrderStatusDto;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuHistory;
-import kitchenpos.domain.menu.repository.MenuHistoryRepository;
-import kitchenpos.domain.menu.repository.MenuRepository;
+import kitchenpos.repository.menu.MenuHistoryRepository;
+import kitchenpos.repository.menu.MenuRepository;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
-import kitchenpos.domain.order.repository.OrderLineItemRepository;
-import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.table.repository.OrderTableRepository;
+import kitchenpos.repository.order.OrderLineItemRepository;
+import kitchenpos.repository.order.OrderRepository;
+import kitchenpos.repository.table.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
