+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -3,9 +3,9 @@ package kitchenpos.application;
 import kitchenpos.application.dto.request.CreateTableGroupDto;
 import kitchenpos.application.dto.response.TableGroupDto;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.table.repository.OrderTableRepository;
-import kitchenpos.domain.table.repository.TableGroupRepository;
+import kitchenpos.repository.order.OrderRepository;
+import kitchenpos.repository.table.OrderTableRepository;
+import kitchenpos.repository.table.TableGroupRepository;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTables;
 import kitchenpos.domain.table.TableGroup;
