+++ src/main/java/kitchenpos/application/TableService.java
@@ -6,8 +6,8 @@ import kitchenpos.application.dto.request.EmptyTableDto;
 import kitchenpos.application.dto.response.TableDto;
 import kitchenpos.application.dto.request.UpdateGuestNumberDto;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.table.repository.OrderTableRepository;
+import kitchenpos.repository.order.OrderRepository;
+import kitchenpos.repository.table.OrderTableRepository;
 import kitchenpos.domain.table.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
