+++ src/main/java/kitchenpos/repository/menu/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.menu.repository;
+package kitchenpos.repository.menu;
 
 import java.util.List;
 import kitchenpos.domain.menu.MenuGroup;
