+++ src/main/java/kitchenpos/repository/table/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.table.repository;
+package kitchenpos.repository.table;
 
 import java.util.List;
 import kitchenpos.domain.table.TableGroup;
